requirements.txt
dj-database-url==2.2.0
Django==5.0.4
django-heroku==0.3.1
djangorestframework==3.15.1
ipython==8.23.0
langchain==0.1.15
langchain-community==0.0.32
langchain-core==0.1.42
langchain-openai==0.1.3
langchain-text-splitters==0.0.1
onnxruntime==1.17.1
openai==1.16.2
pymongo==4.7.3
pyparsing==3.1.2
pypdf==4.2.0
PyPDF2==3.0.1
requests==2.31.0
gunicorn==22.0.0